Added all necessary elements for functioning quiz
Completed base text elements and JavaScript logic for a functioning quiz. Was able to get the timer to work and the elements to properly display and react to the user's input. Added a separate JavaScript and HTML file for the high scores and allowed a path back to the main game.